Fixed a few snapshot bugs
