+++ Source/EventFlow.Tests/EventFlow.Tests.csproj
@@ -58,7 +58,6 @@
     <Compile Include="IntegrationTests\ReadStores\ReadModels\InMemoryThingyReadModel.cs" />
     <Compile Include="IntegrationTests\ResolverTests.cs" />
     <Compile Include="IntegrationTests\SnapshotStores\InMemorySnapshotStoreTests.cs" />
-    <Compile Include="IntegrationTests\SnapshotTests.cs" />
     <Compile Include="UnitTests\Aggregates\AggregateRootNameTests.cs" />
     <Compile Include="UnitTests\Aggregates\AggregateRootApplyEventTests.cs" />
     <Compile Include="UnitTests\Aggregates\AggregateStateTests.cs" />
